Added tests for consumer side
