+++ client/src/main/java/org/apache/rocketmq/client/trace/hook/micrometer/RocketMqSenderObservationConvention.java
@@ -19,6 +19,9 @@ package org.apache.rocketmq.client.trace.hook.micrometer;
 import io.micrometer.observation.Observation.Context;
 import io.micrometer.observation.ObservationConvention;
 
+/**
+ * An {@link ObservationConvention} for {@link RocketMqSenderContext}.
+ */
 public interface RocketMqSenderObservationConvention extends ObservationConvention<RocketMqSenderContext> {
 
     @Override
