+++ client/src/test/java/org/apache/rocketmq/client/trace/DefaultMQConsumerWithOpenTracingTest.java
@@ -20,220 +20,30 @@ package org.apache.rocketmq.client.trace;
 import io.opentracing.mock.MockSpan;
 import io.opentracing.mock.MockTracer;
 import io.opentracing.tag.Tags;
-import java.io.ByteArrayOutputStream;
-import java.net.InetSocketAddress;
-import java.util.Collections;
-import java.util.HashSet;
-import java.util.List;
-import java.util.Map;
-import java.util.Set;
-import java.util.concurrent.Callable;
-import java.util.concurrent.ConcurrentMap;
-import java.util.concurrent.CountDownLatch;
-import java.util.concurrent.TimeUnit;
-import java.util.concurrent.atomic.AtomicReference;
-import org.apache.commons.lang3.reflect.FieldUtils;
-import org.apache.rocketmq.client.consumer.DefaultMQPushConsumer;
-import org.apache.rocketmq.client.consumer.PullCallback;
-import org.apache.rocketmq.client.consumer.PullResult;
-import org.apache.rocketmq.client.consumer.PullStatus;
-import org.apache.rocketmq.client.consumer.listener.ConsumeConcurrentlyContext;
-import org.apache.rocketmq.client.consumer.listener.ConsumeConcurrentlyStatus;
-import org.apache.rocketmq.client.consumer.listener.MessageListenerConcurrently;
-import org.apache.rocketmq.client.consumer.store.OffsetStore;
-import org.apache.rocketmq.client.consumer.store.ReadOffsetType;
-import org.apache.rocketmq.client.exception.MQBrokerException;
-import org.apache.rocketmq.client.exception.MQClientException;
-import org.apache.rocketmq.client.impl.CommunicationMode;
-import org.apache.rocketmq.client.impl.FindBrokerResult;
-import org.apache.rocketmq.client.impl.MQClientAPIImpl;
-import org.apache.rocketmq.client.impl.MQClientManager;
-import org.apache.rocketmq.client.impl.consumer.ConsumeMessageConcurrentlyService;
-import org.apache.rocketmq.client.impl.consumer.DefaultMQPushConsumerImpl;
-import org.apache.rocketmq.client.impl.consumer.ProcessQueue;
-import org.apache.rocketmq.client.impl.consumer.PullAPIWrapper;
-import org.apache.rocketmq.client.impl.consumer.PullMessageService;
-import org.apache.rocketmq.client.impl.consumer.PullRequest;
-import org.apache.rocketmq.client.impl.consumer.PullResultExt;
-import org.apache.rocketmq.client.impl.consumer.RebalancePushImpl;
-import org.apache.rocketmq.client.impl.factory.MQClientInstance;
+import org.apache.rocketmq.client.hook.ConsumeMessageHook;
 import org.apache.rocketmq.client.trace.hook.ConsumeMessageOpenTracingHookImpl;
-import org.apache.rocketmq.common.message.MessageClientExt;
-import org.apache.rocketmq.common.message.MessageDecoder;
-import org.apache.rocketmq.common.message.MessageExt;
-import org.apache.rocketmq.common.message.MessageQueue;
-import org.apache.rocketmq.remoting.RPCHook;
-import org.apache.rocketmq.remoting.exception.RemotingException;
-import org.apache.rocketmq.remoting.protocol.header.PullMessageRequestHeader;
-import org.junit.After;
-import org.junit.Before;
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.mockito.Mock;
-import org.mockito.Mockito;
-import org.mockito.invocation.InvocationOnMock;
-import org.mockito.junit.MockitoJUnitRunner;
-import org.mockito.stubbing.Answer;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.awaitility.Awaitility.waitAtMost;
-import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.ArgumentMatchers.anyBoolean;
-import static org.mockito.ArgumentMatchers.anyLong;
-import static org.mockito.ArgumentMatchers.anyString;
-import static org.mockito.ArgumentMatchers.nullable;
-import static org.mockito.Mockito.doReturn;
-import static org.mockito.Mockito.spy;
-import static org.mockito.Mockito.when;
 
-@RunWith(MockitoJUnitRunner.class)
-public class DefaultMQConsumerWithOpenTracingTest {
-    private String consumerGroup;
+public class DefaultMQConsumerWithOpenTracingTest extends AbstractDefaultMQConsumerTest {
 
-    private String topic = "FooBar";
-    private String brokerName = "BrokerA";
-    private MQClientInstance mQClientFactory;
-
-    @Mock
-    private MQClientAPIImpl mQClientAPIImpl;
-    private PullAPIWrapper pullAPIWrapper;
-    private RebalancePushImpl rebalancePushImpl;
-    private DefaultMQPushConsumer pushConsumer;
     private final MockTracer tracer = new MockTracer();
 
-    @Before
-    public void init() throws Exception {
-        ConcurrentMap<String, MQClientInstance> factoryTable = (ConcurrentMap<String, MQClientInstance>) FieldUtils.readDeclaredField(MQClientManager.getInstance(), "factoryTable", true);
-        for (Map.Entry<String, MQClientInstance> entry : factoryTable.entrySet()) {
-            entry.getValue().shutdown();
-        }
-        factoryTable.clear();
-
-        when(mQClientAPIImpl.pullMessage(anyString(), any(PullMessageRequestHeader.class),
-            anyLong(), any(CommunicationMode.class), nullable(PullCallback.class)))
-            .thenAnswer(new Answer<PullResult>() {
-                @Override
-                public PullResult answer(InvocationOnMock mock) throws Throwable {
-                    PullMessageRequestHeader requestHeader = mock.getArgument(1);
-                    MessageClientExt messageClientExt = new MessageClientExt();
-                    messageClientExt.setTopic(topic);
-                    messageClientExt.setQueueId(0);
-                    messageClientExt.setMsgId("123");
-                    messageClientExt.setBody(new byte[] {'a'});
-                    messageClientExt.setOffsetMsgId("234");
-                    messageClientExt.setBornHost(new InetSocketAddress(8080));
-                    messageClientExt.setStoreHost(new InetSocketAddress(8080));
-                    PullResult pullResult = createPullResult(requestHeader, PullStatus.FOUND, Collections.<MessageExt>singletonList(messageClientExt));
-                    ((PullCallback) mock.getArgument(4)).onSuccess(pullResult);
-                    return pullResult;
-                }
-            });
-
-        consumerGroup = "FooBarGroup" + System.currentTimeMillis();
-        pushConsumer = new DefaultMQPushConsumer(consumerGroup);
-        pushConsumer.getDefaultMQPushConsumerImpl().registerConsumeMessageHook(
-            new ConsumeMessageOpenTracingHookImpl(tracer));
-        pushConsumer.setNamesrvAddr("127.0.0.1:9876");
-        pushConsumer.setPullInterval(60 * 1000);
-
-        OffsetStore offsetStore = Mockito.mock(OffsetStore.class);
-        Mockito.when(offsetStore.readOffset(any(MessageQueue.class), any(ReadOffsetType.class))).thenReturn(0L);
-        pushConsumer.setOffsetStore(offsetStore);
-
-        pushConsumer.registerMessageListener(new MessageListenerConcurrently() {
-            @Override
-            public ConsumeConcurrentlyStatus consumeMessage(List<MessageExt> msgs,
-                ConsumeConcurrentlyContext context) {
-                return null;
-            }
-        });
-
-        DefaultMQPushConsumerImpl pushConsumerImpl = pushConsumer.getDefaultMQPushConsumerImpl();
-
-        // suppress updateTopicRouteInfoFromNameServer
-        pushConsumer.changeInstanceNameToPID();
-        mQClientFactory = MQClientManager.getInstance().getOrCreateMQClientInstance(pushConsumer, (RPCHook) FieldUtils.readDeclaredField(pushConsumerImpl, "rpcHook", true));
-        FieldUtils.writeDeclaredField(mQClientFactory, "mQClientAPIImpl", mQClientAPIImpl, true);
-        mQClientFactory = spy(mQClientFactory);
-        factoryTable.put(pushConsumer.buildMQClientId(), mQClientFactory);
-        doReturn(false).when(mQClientFactory).updateTopicRouteInfoFromNameServer(anyString());
-
-        doReturn(new FindBrokerResult("127.0.0.1:10911", false)).when(mQClientFactory).findBrokerAddressInSubscribe(anyString(), anyLong(), anyBoolean());
-
-        Set<MessageQueue> messageQueueSet = new HashSet<>();
-        messageQueueSet.add(createPullRequest().getMessageQueue());
-        pushConsumerImpl.updateTopicSubscribeInfo(topic, messageQueueSet);
-
-        pushConsumer.subscribe(topic, "*");
-        pushConsumer.start();
+    @Override
+    ConsumeMessageHook consumeMessageHook() {
+        return new ConsumeMessageOpenTracingHookImpl(tracer);
     }
 
-    @After
-    public void terminate() {
-        pushConsumer.shutdown();
+    @Override
+    boolean waitCondition() {
+        return tracer.finishedSpans().size() == 1;
     }
 
-    @Test
-    public void testPullMessage_WithTrace_Success() throws InterruptedException, RemotingException, MQBrokerException, MQClientException {
-        final CountDownLatch countDownLatch = new CountDownLatch(1);
-        final AtomicReference<MessageExt> messageAtomic = new AtomicReference<>();
-        pushConsumer.getDefaultMQPushConsumerImpl().setConsumeMessageService(new ConsumeMessageConcurrentlyService(pushConsumer.getDefaultMQPushConsumerImpl(), new MessageListenerConcurrently() {
-            @Override
-            public ConsumeConcurrentlyStatus consumeMessage(List<MessageExt> msgs,
-                ConsumeConcurrentlyContext context) {
-                messageAtomic.set(msgs.get(0));
-                countDownLatch.countDown();
-                return ConsumeConcurrentlyStatus.CONSUME_SUCCESS;
-            }
-        }));
-
-        PullMessageService pullMessageService = mQClientFactory.getPullMessageService();
-        pullMessageService.executePullRequestImmediately(createPullRequest());
-        countDownLatch.await(30, TimeUnit.SECONDS);
-        MessageExt msg = messageAtomic.get();
-        assertThat(msg).isNotNull();
-        assertThat(msg.getTopic()).isEqualTo(topic);
-        assertThat(msg.getBody()).isEqualTo(new byte[] {'a'});
-
-        // wait until consumeMessageAfter hook of tracer is done surely.
-        waitAtMost(1, TimeUnit.SECONDS).until(new Callable() {
-            @Override
-            public Object call() throws Exception {
-                return tracer.finishedSpans().size() == 1;
-            }
-        });
-
+    @Override
+    void assertResults() {
         MockSpan span = tracer.finishedSpans().get(0);
         assertThat(span.tags().get(Tags.MESSAGE_BUS_DESTINATION.getKey())).isEqualTo(topic);
         assertThat(span.tags().get(Tags.SPAN_KIND.getKey())).isEqualTo(Tags.SPAN_KIND_CONSUMER);
         assertThat(span.tags().get(TraceConstants.ROCKETMQ_SUCCESS)).isEqualTo(true);
     }
-
-    private PullRequest createPullRequest() {
-        PullRequest pullRequest = new PullRequest();
-        pullRequest.setConsumerGroup(consumerGroup);
-        pullRequest.setNextOffset(1024);
-
-        MessageQueue messageQueue = new MessageQueue();
-        messageQueue.setBrokerName(brokerName);
-        messageQueue.setQueueId(0);
-        messageQueue.setTopic(topic);
-        pullRequest.setMessageQueue(messageQueue);
-        ProcessQueue processQueue = new ProcessQueue();
-        processQueue.setLocked(true);
-        processQueue.setLastLockTimestamp(System.currentTimeMillis());
-        pullRequest.setProcessQueue(processQueue);
-
-        return pullRequest;
-    }
-
-    private PullResultExt createPullResult(PullMessageRequestHeader requestHeader, PullStatus pullStatus,
-        List<MessageExt> messageExtList) throws Exception {
-        ByteArrayOutputStream outputStream = new ByteArrayOutputStream();
-        for (MessageExt messageExt : messageExtList) {
-            outputStream.write(MessageDecoder.encode(messageExt, false));
-        }
-        return new PullResultExt(pullStatus, requestHeader.getQueueOffset() + messageExtList.size(), 123, 2048, messageExtList, 0, outputStream.toByteArray());
-    }
-
 }
